generalised some components & laid the groundwork for more tools
