+++ src/index.css
@@ -1,5 +1,7 @@
 @import url("https://fonts.googleapis.com/css?family=Nunito:400,700,900");
 
+/** LAYOUT */
+
 .main-body {
   min-height: 100vh;
   display: flex;
@@ -14,7 +16,8 @@
   max-width: 920px !important;
 }
 
-/** override bulma default to make modal header look good */
+/** BULMA OVERRIDES */
+
 .modal-card-title {
   margin-bottom: 0 !important;
 }
@@ -22,6 +25,8 @@
 .message-header {
   margin-bottom: 0 !important;
 }
+
+/** LOGOS */
 
 .logo-icon {
   width: 40px;
@@ -47,6 +52,8 @@
   font-family: "Nunito", monospace;
 }
 
+/** ELEMENTS */
+
 .simple-card {
   box-shadow: none;
   background: #fcfcfc !important;
@@ -65,8 +72,124 @@ a.underlined-style {
   color: inherit !important;
 }
 
+/** PAGES */
+
 .light-border {
   border: 1px solid hsl(0, 0%, 71%); /* has-text-grey-light */
   border-radius: 10px;
   padding: 10px;
+}
+
+/** PAGE RESULT */
+
+.review-content {
+  border: 1px solid hsl(0, 0%, 71%); /* has-text-grey-light */
+  border-radius: 10px;
+}
+
+.review-content label {
+  font-family: monospace;
+  margin-right: 10px;
+  font-size: medium;
+}
+
+.review-content input {
+  font-family: monospace;
+  border: 0;
+  padding: 0;
+  background-color: lightgreen;
+}
+
+.review-content input::placeholder {
+  background-color: hsl(48, 100%, 67%);
+}
+
+.review-content input:focus {
+  outline: 0 !important;
+}
+
+.review-content select {
+  background-color: lightgreen;
+  font-family: monospace;
+  border: 0;
+}
+
+/** auto text area */
+
+.autotext-area {
+  width: 100%;
+  resize: none;
+  background: transparent;
+  border-radius: 0;
+  font-size: medium;
+  font-family: monospace;
+  padding: 0;
+  margin: 0;
+  scroll-behavior: none;
+  margin-bottom: 0px !important;
+  line-height: 1.5;
+  border: none;
+}
+
+.autotext-area:focus {
+  outline: 0 !important;
+}
+
+.autotext-area:hover {
+  border-color: #c6c6c6;
+  color: hsl(217, 71%, 53%);
+}
+
+/** results */
+
+.results-container {
+  width: 100%;
+}
+
+.big-div {
+  flex-grow: 1;
+}
+
+.plus-button-holder {
+  text-align: center;
+  margin-top: -13px;
+}
+
+.plus-button {
+  width: 26px;
+  height: 26px;
+  font-size: xx-small !important;
+}
+
+.top-content {
+  display: flex;
+  border-top: 1px solid lightgrey;
+  border-left: 1px solid lightgrey;
+  border-right: 1px solid lightgrey;
+  border-bottom: 1px dashed lightgrey;
+  border-radius: 5px 5px 0px 0px;
+}
+
+.normal-content {
+  display: flex;
+  border-left: 1px solid lightgrey;
+  border-right: 1px solid lightgrey;
+  margin-top: -13px;
+  border-bottom: 1px dashed lightgrey;
+}
+
+.bottom-content {
+  display: flex;
+  border-left: 1px solid lightgrey;
+  border-right: 1px solid lightgrey;
+  margin-top: -13px;
+  border-bottom: 1px solid lightgrey;
+  border-radius: 0px 0px 5px 5px;
+}
+
+/** UTILITY **/
+
+.horizontal-line {
+  height: 0px;
+  border-bottom: 1px dashed hsl(0, 0%, 71%); /* has-text-grey-light */
 }
