--- src/pages/performance-review/perfreview.css
@@ -1,111 +1,0 @@
-.review-content {
-  border: 1px solid hsl(0, 0%, 71%); /* has-text-grey-light */
-  border-radius: 10px;
-}
-
-.review-content label {
-  font-family: monospace;
-  margin-right: 10px;
-  font-size: medium;
-}
-
-.review-content input {
-  font-family: monospace;
-  border: 0;
-  padding: 0;
-  background-color: lightgreen;
-}
-
-.review-content input::placeholder {
-  background-color: hsl(48, 100%, 67%);
-}
-
-.review-content input:focus {
-  outline: 0 !important;
-}
-
-.review-content select {
-  background-color: lightgreen;
-  font-family: monospace;
-  border: 0;
-}
-
-/** auto text area */
-
-.autotext-area {
-  width: 100%;
-  resize: none;
-  background: transparent;
-  border-radius: 0;
-  font-size: medium;
-  font-family: monospace;
-  padding: 0;
-  margin: 0;
-  scroll-behavior: none;
-  margin-bottom: 0px !important;
-  line-height: 1.5;
-  border: none;
-}
-
-.autotext-area:focus {
-  outline: 0 !important;
-}
-
-.autotext-area:hover {
-  border-color: #c6c6c6;
-  color: hsl(217, 71%, 53%);
-}
-
-/** results */
-
-.results-container {
-  width: 100%;
-}
-
-.big-div {
-  flex-grow: 1;
-}
-
-.plus-button-holder {
-  text-align: center;
-  margin-top: -13px;
-}
-
-.plus-button {
-  width: 26px;
-  height: 26px;
-  font-size: xx-small !important;
-}
-
-.top-content {
-  display: flex;
-  border-top: 1px solid lightgrey;
-  border-left: 1px solid lightgrey;
-  border-right: 1px solid lightgrey;
-  border-bottom: 1px dashed lightgrey;
-  border-radius: 5px 5px 0px 0px;
-}
-
-.normal-content {
-  display: flex;
-  border-left: 1px solid lightgrey;
-  border-right: 1px solid lightgrey;
-  margin-top: -13px;
-  border-bottom: 1px dashed lightgrey;
-}
-
-.bottom-content {
-  display: flex;
-  border-left: 1px solid lightgrey;
-  border-right: 1px solid lightgrey;
-  margin-top: -13px;
-  border-bottom: 1px solid lightgrey;
-  border-radius: 0px 0px 5px 5px;
-}
-
-/** UTILITY **/
-
-.horizontal-line {
-  height: 0px;
-  border-bottom: 1px dashed hsl(0, 0%, 71%); /* has-text-grey-light */
-}
